python/connect_with_playwright_to_chrome/requirements.txt
kameleo.local_api_client
playwright